test_phuma_code.py
#!/usr/bin/env python3
"""
PHUMA 代码验证脚本
测试各个模块的基本功能
"""

import sys
import os
sys.path.insert(0, 'src')

print("=" * 60)
print("PHUMA 代码验证脚本")
print("=" * 60)
print()

# 测试 1: 导入所有必要的模块
print("📦 测试 1: 导入模块...")
try:
    import torch
    import numpy as np
    import smplx
    import mujoco
    import yaml
    from utils.smpl import load_motion_parms, find_robust_ground
    from utils.robot import HumanoidRetargetKeypoint
    print("✓ 所有核心模块导入成功")
    print(f"  - PyTorch: {torch.__version__}")
    print(f"  - NumPy: {np.__version__}")
    print(f"  - MuJoCo: {mujoco.__version__}")
except Exception as e:
    print(f"✗ 模块导入失败: {e}")
    sys.exit(1)

print()

# 测试 2: 检查项目结构
print("📁 测试 2: 检查项目结构...")
required_dirs = [
    'asset/human_model',
    'asset/humanoid_model/g1',
    'asset/humanoid_model/h1_2',
    'data/human_pose/example',
    'src/curation',
    'src/retarget',
    'src/utils'
]

all_exist = True
for dir_path in required_dirs:
    if os.path.exists(dir_path):
        print(f"✓ {dir_path}")
    else:
        print(f"✗ {dir_path} 不存在")
        all_exist = False

if not all_exist:
    print("⚠️  某些目录缺失，但可以继续")
print()

# 测试 3: 检查示例数据
print("📊 测试 3: 检查示例数据...")
try:
    example_data = np.load('data/human_pose/example/kick.npy')
    print(f"✓ 示例数据加载成功")
    print(f"  - 形状: {example_data.shape}")
    print(f"  - 类型: {example_data.dtype}")
    print(f"  - 帧数: {example_data.shape[0]}")
    print(f"  - 参数维度: {example_data.shape[1]}")

    if example_data.shape[1] == 69:
        print("✓ 数据格式正确 (69 = 3 transl + 3 global_orient + 63 body_pose)")
    else:
        print(f"⚠️  数据维度为 {example_data.shape[1]}，预期为 69")
except Exception as e:
    print(f"✗ 示例数据加载失败: {e}")

print()

# 测试 4: 检查机器人配置
print("🤖 测试 4: 检查机器人配置...")
robot_configs = {
    'g1': 'asset/humanoid_model/g1/config.yaml',
    'h1_2': 'asset/humanoid_model/h1_2/config.yaml'
}

for robot_name, config_path in robot_configs.items():
    try:
        if os.path.exists(config_path):
            with open(config_path, 'r') as f:
                config = yaml.safe_load(f)
            print(f"✓ {robot_name} 配置加载成功")
            print(f"  - 身体数量: {len(config.get('body_names', []))}")
            print(f"  - 关键点数量: {len(config.get('keypoints', []))}")
        else:
            print(f"✗ {robot_name} 配置文件不存在")
    except Exception as e:
        print(f"✗ {robot_name} 配置加载失败: {e}")

print()

# 测试 5: 检查 MuJoCo 模型
print("🎮 测试 5: 检查 MuJoCo 模型...")
mujoco_models = {
    'g1': 'asset/humanoid_model/g1/scene.xml',
    'h1_2': 'asset/humanoid_model/h1_2/scene.xml'
}

for robot_name, model_path in mujoco_models.items():
    try:
        if os.path.exists(model_path):
            model = mujoco.MjModel.from_xml_path(model_path)
            print(f"✓ {robot_name} MuJoCo 模型加载成功")
            print(f"  - DOF: {model.nv}")
            print(f"  - Bodies: {model.nbody}")
        else:
            print(f"✗ {robot_name} MuJoCo 模型文件不存在")
    except Exception as e:
        print(f"✗ {robot_name} MuJoCo 模型加载失败: {e}")

print()

# 测试 6: 测试工具函数
print("🔧 测试 6: 测试工具函数...")
try:
    # 测试加载运动参数（不需要 SMPL-X 模型）
    test_data = np.random.randn(10, 69).astype(np.float32)
    np.save('/tmp/test_motion.npy', test_data)

    motion_params = load_motion_parms(
        '/tmp/test_motion.npy', foot_contact=False)
    print(f"✓ load_motion_parms 函数工作正常")
    print(f"  - transl 形状: {motion_params['transl'].shape}")
    print(f"  - global_orient 形状: {motion_params['global_orient'].shape}")
    print(f"  - body_pose 形状: {motion_params['body_pose'].shape}")

    os.remove('/tmp/test_motion.npy')
except Exception as e:
    print(f"✗ 工具函数测试失败: {e}")

print()

# 测试 7: SMPL-X 模型检查
print("👤 测试 7: SMPL-X 模型检查...")
smplx_dir = 'asset/human_model/smplx'
if os.path.exists(smplx_dir):
    smplx_files = os.listdir(smplx_dir)
    required_files = ['SMPLX_NEUTRAL.npz',
                      'SMPLX_MALE.npz', 'SMPLX_FEMALE.npz']

    if len(smplx_files) == 0:
        print(f"⚠️  SMPL-X 目录为空")
        print(f"   请从 https://smpl-x.is.tue.mpg.de/ 下载模型文件")
        print(f"   需要的文件: {', '.join(required_files)}")
    else:
        print(f"✓ SMPL-X 目录存在，包含 {len(smplx_files)} 个文件")
        for required in required_files:
            if required in smplx_files:
                print(f"  ✓ {required}")
            else:
                print(f"  ✗ {required} (缺失)")
else:
    print(f"✗ SMPL-X 目录不存在")

print()

# 测试总结
print("=" * 60)
print("📋 测试总结")
print("=" * 60)
print("✓ 基础环境配置正确")
print("✓ 代码模块可以正常导入")
print("✓ 示例数据格式正确")
print("✓ 机器人配置文件完整")
print()
print("⚠️  注意事项:")
print("1. SMPL-X 模型文件需要手动下载")
print("2. 下载地址: https://smpl-x.is.tue.mpg.de/")
print("3. 放置位置: asset/human_model/smplx/")
print()
print("🎉 代码验证完成!")
print("=" * 60)
